% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\alias{Caenorhabditis_elegans}
\alias{Danio_rerio}
\alias{Drosophila_melanogaster}
\alias{Homo_sapiens}
\alias{Mus_musculus}
\alias{Rattus_norvegicus}
\title{.onLoad Function for Resource Name}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The name of the Library.}

\item{pkgname}{The name of the Package.}
}
\value{
a set of functions specific to each resource species
}
\description{
Resource Name Function
}
\examples{
CoSIAdata::Danio_rerio()
}
\references{
Morgan M, Shepherd L (2022). ExperimentHub: Client to access ExperimentHub resources. R package version 2.6.0.
}
