% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIAdata-package.R
\docType{package}
\name{CoSIAdata}
\alias{CoSIAdata}
\alias{_PACKAGE}
\title{CoSIAdata: VST normalized RNA-Sequencing data with annotations for multiple species samples from Bgee}
\description{
Variance Stabilized Transformation of Read Counts derieved from 
Bgee RNA-Seq Expression Data. Expression Data includes annotations and is 
across 6 species (Homo sapiens, Mus musculus, Rattus norvegicius,Danio rerio,
Drosophila melanogaster, and Caenorhabditis elegans) and across more than 132
tissues. The data is represented as a RData files and is available in 
ExperimentHub.
}
\seealso{
Useful links:
\itemize{
  \item \url{}
}

}
\keyword{internal}
