% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Resource Name Function}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The name of the Library.}

\item{pkgname}{The name of the Package.}
}
\value{
a set of functions specific to each resource species
}
\description{
Resource Name Function
}
\examples{
CoSIAdata::`Danio rerio`()

}
